+++ CHANGELOG.md
@@ -12,6 +12,9 @@
   * Honor the `form` attribute when creating hidden checkbox input
   * Use `to_iso8601` as the standard implementation for safe dates and times
 
+* Deprecations
+  * `form_for` without an anonymous function has been deprecated. v3.0 has deprecated the usage, v3.1 will emit warnings, and v3.2 will fully remove the functionality
+
 * Backwards incompatible changes
   * Strings given as attributes in `tag` and `content_tag` are now emitted as is (without being dasherized) and are also HTML escaped
   * Prefix form id on inputs when it is given to `form_for/3`